Use string resources, more previews and tests adjustments
