+++ app/src/main/res/values/strings.xml
@@ -270,4 +270,14 @@
     <string name="onboarding_page_content_importer">There is a vocabulary importer that accepts .json files in a specific format.</string>
     <string name="onboarding_page_content_importer_extended">You can either import words from the settings or by clicking on a .json file and selecting the app.</string>
 
+    <!-- # VocabularyDetailScreen # -->
+    <string name="vocabulary_detail_section_pronunciation">Pronunciation</string>
+    <string name="vocabulary_detail_section_notes">Notes</string>
+    <string name="vocabulary_detail_section_metadata">Metadata</string>
+    <string name="vocabulary_detail_section_metadata_created">Created: %1$s</string>
+    <string name="vocabulary_detail_section_metadata_edited">Last edited: %1$s</string>
+    <string name="vocabulary_detail_chip_group">%1$s (%2$s)</string>
+    <string name="vocabulary_detail_chip_gender">Gender: %1$s (%2$s)</string>
+    <string name="vocabulary_detail_chip_endings">Endings: %1$s</string>
+    <string name="vocabulary_detail_edit_btn">Edit</string>
 </resources>
